update
fix bugs and add PG60

+++ ATmega32U4/usb_keyboard.c
@@ -395,7 +395,7 @@ ISR(USB_GEN_vect)
 					keyboard_buffer.keyboard_idle_count = 0;
 					UEDATX=keyboard_report.modifier;
 					UEDATX=keyboard_report.reserved;
-					for (int i=0; i<6; i++) {
+					for (uint8_t i=0; i<6; i++) {
 						UEDATX = keyboard_report.keycode[i];
 					}
 					ReleaseTX();
